refactor: Name 내에 필드 추가로 인한 사용자 클래스명 변경 Name -> User
